[Issue-0000] Alternative Loading state refactor
[Summary]
The Loading UI is usually shown in conjunction with other UI, such as an
error view, or valid data. Consuming it with a when statement and
treating it like any other state may not make sense with Compose and
Swift UI. I tried a different approach here, giving the Loading state
the last valid DataState as well:
https://github.com/touchlab-lab/KaMPKit/pull/6

[Fix]
In this PR, I'm using the classic NetworkBoundResource approach, where
the DataState is not a sealed class; instead, it contains nullable
versions of all possible states.

[Testing]
- `./gradlew :app:build`
- `./gradlew :shared:build`
- `xcodebuild -workspace ios/KaMPKitiOS.xcworkspace -scheme KaMPKitiOS
    -sdk iphoneos -configuration Debug build -destination name="iPhone 8"`
- manual testing