Wraps `runTest`'s `block` with `coroutineScope`
There is a problem that if you call `launch` inside of `runTest`, `runTest` will leave the scope without waiting for `launch` to complete.
In particular, on iOS, this can cause unexpected crashes since you will `launch` inside a `GlobalScope`.
By wrapping it with `coroutineScope`, the `runTest` will wait for the `launch`.